renamed chroma adapter
