+++ hri/packages/embeddings/scripts/knowledge_base.py
@@ -2,7 +2,7 @@
 import rclpy
 from rclpy.node import Node
 from frida_interfaces.srv import AnswerQuestion
-from ChromaAdapter import ChromaAdapter
+from hri.packages.embeddings.scripts.chroma_adapter import ChromaAdapter
 from sentence_transformers import SentenceTransformer
 from llama_cpp import Llama
